getStore() -> store
