--- code/addons/test/src/universal-store/manager.ts
@@ -1,13 +1,0 @@
-import { experimental_UniversalStore } from 'storybook/internal/manager-api';
-
-import {
-  type UniversalStoreEvent,
-  type UniversalStoreState,
-  universalStoreConfig,
-} from '../constants';
-
-export const getStore = (leader = false) =>
-  experimental_UniversalStore.create<UniversalStoreState, UniversalStoreEvent>({
-    ...universalStoreConfig,
-    leader,
-  });
